reorganization of utils into Utils directory
