+++ components/src/Utils/index.js
@@ -1,0 +1,3 @@
+export { default as InputClickableArea } from "./InputClickableArea";
+export { default as omit } from "./omit";
+export { default as subPx } from "./subPx";
