+++ components/src/Utils/omit.js
@@ -1,0 +1,7 @@
+const omit = (obj, prop) => {
+  const res = Object.assign({}, obj);
+  delete res[prop];
+  return res;
+};
+
+export default omit;
